requirements.txt
numpy
Pillow
opencv-python
imutils
tensorflow
tflearn
scikit-learn